requirements.txt
Pillow
discord
pymongo

